Added the ability to configure the flash CSS class.
Adding an application setting such as...

<appSettings>
  <add key="FlashCssClass" value="flash"/>
</appSettings>

will allow you to configure the name of the CSS class that is applied
to flash messages. I was inspired by Twitter's bootstrap project, which
supports a messaging style, but uses the CSS class "alert-message". As
it turns out, this was quite an easy item to configure.